fix: extract shared sentry init to eliminate 102-line duplication (SonarCloud)
SonarCloud flagged 9.6% duplication on new code. Root cause: 3 sentry-instrument.ts
files (shared, core-api, gateway) contained identical helper functions:
- findUp (19 lines)
- readTextSafe (7 lines)
- resolveGitSha (25 lines)
- resolveRepoVersion (9 lines)
- getSentryRelease (35 lines)

Extracted into backend/erxes-api-shared/src/utils/sentry-init.ts:
- initErxesSentry(beforeSend?, serverName?) function
- getSentryRelease() now lives in the shared module

Each sentry-instrument.ts file is now ~5 lines:
  import { initErxesSentry, sentryExpectedErrorFilter } from 'erxes-api-shared/utils';
  initErxesSentry(sentryExpectedErrorFilter, 'core-api');

Also exported from utils/index.ts barrel.
