+++ backend/erxes-api-shared/src/utils/index.ts
@@ -22,3 +22,4 @@ export * from './string';
 export * from './bulkUtils';
 export * from './editor';
 export * from './errorClassifier';
+export * from './sentry-init';
